Edge case handling for when file birthtime is 0
In some operating systems and file systems, the file stat returns 0 for the file's creation time.

In this case it's impossible to incrementally index them correctly because breadboard keeps a checkpoint of the last indexed file's birth time, and starts from there. In this case because every file has 0 for their birth time, this doesn't work.

For this edge case, just index the mtime (modified time).

This won't solve the problem fundamentally because it will display modified time instead of created time (files get modified everytime it's touched, including when tagged or favorited), however it will at least be able to incrementally index them because now it has some sort of timestamp to use for checkpoints.