Synchronize per folder
One of the biggest annoyances has been that every time I want to reindex something (perhaps because there was a messed up metatata indexing issue and a new version of breadboard fixed it, and now need to re-index), I have to reindex the whole thing.

Now you can simply reindex a single individual folder by clicking the "reindex" button next to the delete button in the settings page.