Reindex must completely reboot the DB
The whole point of reindexing is to completely repopulate the DB from scratch regardless of which version of the app is being used.

This is important as the app will add more features which may add more attributes to the database (such as the "width" and "height" for this release).

Previously it was not possible to add additional columns to the DB because it had to use the existing schema and reindexing simply "cleared the tables" instead of deleting the DB and starting from scratch with the new schema.

Now with this update, reindexing will:

1. Back up the columns that cannot be synchronized from files (favorites, settings, folders)
2. Delete the DB entirely
3. Recover the DB columns (favorites, settings, folders) from the backup
4. Delete the backup db
5. Synchronize

This will ensure that the DB can keep evolving without being constrained by the old schema going forward.