Fix the dragging problem
Previously, when dragging items outside of the browser, or when dragging items and canceling, the browser will start thinking it's still dragging, and will mess up the dragging interface and the only way to get out of it was to cancel everything.

By stopping the drag engine before initiating the drag and using the native dragging, and then resuming the dragging engine when mouseout is triggered, it works now.